deconvolution + pooling
