+++ src/core.jl
@@ -6,19 +6,15 @@ const Cptr = Ptr{Void}
 macro gs(); if false; esc(:(ccall(("cudaDeviceSynchronize","libcudart"),UInt32,()))); end; end
 
 include("gpu.jl")
+include("common.jl")
 include("conv.jl")
-
-
-
-
+include("pool.jl")
 
 # See if we have a gpu at initialization:
 function __init__()
     try
-        r = gpu(true)
-        # info(r >= 0 ? "Knet using GPU $r" : "No GPU found, Knet using the CPU")
+        r = gpu(true)        
     catch e
         gpu(false)
-        # warn("Knet using the CPU: $e")
     end
 end
